Fixed tiny bugs

+++ lib/Screens/ProfilePage.dart
@@ -38,14 +38,56 @@ class _ProfilePageState extends State<ProfilePage> {
                 'Profile',
                 actions: [
                   IconButton(
+                      splashRadius: 20,
                       onPressed: () async {
-                        await storage.delete(key: 'jwt');
-                        Navigator.of(context).pushAndRemoveUntil(
-                            MaterialPageRoute(builder: (context) => Login()),
-                            (Route<dynamic> route) => false);
+                        showDialog(
+                          context: context,
+                          builder: (context) => AlertDialog(
+                            backgroundColor: colors.darkScaffoldColor,
+                            title: Text(
+                              "Are you sure?",
+                              style: GoogleFonts.poppins(
+                                  color: colors.darkPrimaryTextColor),
+                            ),
+                            content: Text(
+                              "You will be logged out",
+                              style: GoogleFonts.poppins(
+                                  color: colors.darkPrimaryTextColor),
+                            ),
+                            actions: <Widget>[
+                              TextButton(
+                                child: Text(
+                                  'No',
+                                  style: GoogleFonts.poppins(
+                                      color: colors.darkPrimaryTextColor),
+                                ),
+                                onPressed: () {
+                                  Navigator.of(context).pop();
+                                },
+                              ),
+                              ElevatedButton(
+                                style: ElevatedButton.styleFrom(
+                                    primary: colors.darkAccentColor,
+                                    shape: RoundedRectangleBorder(
+                                        borderRadius: BorderRadius.all(
+                                            Radius.circular(20)))),
+                                child: Text('Yes',
+                                    style: TextStyle(
+                                        color: colors.darkPrimaryTextColor)),
+                                onPressed: () async {
+                                  await storage.delete(key: 'jwt');
+                                  Navigator.of(context).pushAndRemoveUntil(
+                                      MaterialPageRoute(
+                                          builder: (context) => Login()),
+                                      (Route<dynamic> route) => false);
+                                },
+                              ),
+                            ],
+                          ),
+                        );
                       },
                       icon: Icon(
-                        Icons.logout,
+                        Icons.login_outlined,
                         color: colors.darkAccentColor,
                       ))
                 ],
